requirements.txt
pyinstaller
requests
pyjwt
